backend/requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
python-dotenv
psycopg2-binary
